// LDC.Atlas.WebApp/Client/app/shared/entities/section-traffic.entity.ts
export class SectionTraffic {
    sectionId: number;
    blDate: Date;
    blReference: string;
    vesselCode: string;
    companyId: string;
    dataVersionId: number;
    createdDateTime: Date;
    createdBy: string;
    modifiedDateTime: Date;
    modifiedBy: string;
    contractDate: Date;
    portOrigin: string;
    portDestination: string;
    marketSector: string;
    shippingStatusCode: string;
    shippingStatusId: number;
}